Make Chaeri bigger

+++ core/src/main/java/science/skywhale/bloodmortgage/Character.java
@@ -166,7 +166,7 @@ public class Character {
 	}
 	
 	
-	public void render (float delta, SpriteBatch batch, int tilesize, TiledMapTileLayer mapLayer) {
+	public void render (float delta, SpriteBatch batch, float tilesize, TiledMapTileLayer mapLayer) {
 		if (animation != null && (horiSpeed != 0 || vertiSpeed != 0))
 		{
 			elapsedTime += delta;
